🧚‍♀️ Python post method updated

+++ src/main/java/otaku/info/controller/PythonController.java
@@ -57,7 +57,7 @@ public class PythonController {
      * @throws JSONException
      */
 //    public String post(Map<String, String> headers, String json) {
-    public String post(Integer teamId, String text) throws JSONException {
+    public String post(Long teamId, String text) throws JSONException {
 
         // Twitter投稿が終わった後にLINE通知するテキストを詰めていくリスト
         List<String> lineList = new ArrayList<>();
